aromawine3-new_update__with_checkout/wineproject/tokens.py
from rest_framework.authentication import SessionAuthentication
from django.contrib.auth.tokens import PasswordResetTokenGenerator
from six import text_type

class CsrfExemptSessionAuthentication(SessionAuthentication):
    def enforce_csrf(self, request):
        return  # To not perform the csrf check previously happening




class TokenGenerator(PasswordResetTokenGenerator):
    def _make_hash_value(self, user, timestamp):
        return (
            text_type(user.pk) + text_type(timestamp) +
            text_type(user.is_active)
        )
account_activation_token = TokenGenerator()
