aromawine3-new_update__with_checkout/requirements_get.txt
asgiref==3.2.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-autoslug==1.9.7
django-cors-headers==3.5.0
django-date-extensions==3.1
django-filter==2.2.0
django-import-export==2.1.0
django-rest-knox==4.1.0
django-summernote==0.8.11.6
django-webpack-loader==0.7.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
